requirements.txt
markdown>=3.4.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.20.0
